requirements.txt
Flask==3.0.0
pyserial>=3.5
requests>=2.31.0
Pillow>=10.0.0
websocket-client>=1.6.0
svgwrite>=1.4.3
svg.path>=6.0
numpy>=1.24.0
opencv-python>=4.8.0
scikit-image>=0.21.0
